reorganzie cross scenario cuts as an extension + spoke
