+++ examples/uc/cs_uc.py
@@ -16,7 +16,7 @@ from mpisppy.cylinders.xhatlooper_bounder import XhatLooperInnerBound
 # extensions for the hub
 from mpisppy.extensions.extension import MultiExtension
 from mpisppy.extensions.cross_scen_extension import CrossScenarioExtension
-from mpisppy.cylinders.cross_scen_hub import CrossScenarioHub
+from mpisppy.cylinders.hub import PHHub
 from mpisppy.cylinders.cross_scen_spoke import CrossScenarioCutSpoke
 from mpisppy.opt.lshaped import LShapedMethod
 
@@ -121,7 +121,7 @@ if __name__ == "__main__":
 
     # PH hub
     hub_dict = {
-        "hub_class": CrossScenarioHub,
+        "hub_class": PHHub,
         "hub_kwargs": dict(),
         "opt_class": PH,
         "opt_kwargs": {
